Added a playbook for tearing down OpenStack environments
The openstack_teardown playbook will destroy and undefine virtual machines used
to run OpenStack and remove all their storage. This is meant for taking down
devel and possibly test environments, so there is a variable that must
explicitly be set to "no" or the playbook will do nothing. This variable can be
set in the group_vars of devel/test environments.

Some things had to be added to variables for environments:

 - a variable called "protected" that defines whether the new playbook can run
 - the puppetmaster host in the inventory of all environments (this is useful in
general, so also added it to prod inventories)
 - the name of the puppetmaster host defined per environment in vars (might not
be necessary, but i could not figure out how to get the same info through
inventory alone)

It was also necessary to add puppetmasters under ssh.config. I also changed the
rules for devel environments a bit so that their host keys are not stored on the
machine running Ansible. This would not make much sense, as these environments
are so ephemeral anyway.